refactor: move elements from header to navigation
