+++ src/components/Header/Header.css
@@ -10,15 +10,10 @@
   max-width: 1140px;
   margin: 0 auto;
   display: flex;
-  padding: 18px 30px;
+  padding: 0 30px;
   justify-content: space-between;
   align-items: center;
-}
-
-.header__nav {
-  width: 100%;
-  display: flex;
-  align-items: center;
+  height: 74px;
 }
 
 .header__logo {
@@ -32,182 +27,4 @@
 .header__logo:hover {
   cursor: pointer;
   opacity: 0.8;
-}
-
-.header__links {
-  margin: 0;
-  padding: 0;
-  margin-left: 45px;
-  display: flex;
-  gap: 16px;
-  list-style-type: none;
-}
-
-.header__links-item {
-  margin: 0;
-  padding: 0;
-  font-size: 13px;
-  font-weight: 400;
-  line-height: 18px;
-  text-decoration: none;
-  color: #fff;
-}
-
-.header__links-item:hover {
-  opacity: 0.7;
-  cursor: pointer;
-}
-
-.header__acc {
-  display: flex;
-  gap: 7px;
-  text-decoration: none;
-  align-items: center;
-  border-radius: 40px;
-}
-
-.header__acc :hover {
-  opacity: 0.8;
-}
-
-.header__acc_blue {
-  box-shadow: 0px 3px 14px 0px rgba(0, 0, 0, 0.08);
-  background: #073042;
-}
-.header__acc_dark {
-  box-shadow: 0px 3px 14px 0px rgba(0, 0, 0, 0.08);
-  background: #202020;
-}
-
-.header__acc-text {
-  margin: 0;
-  padding: 0;
-  color: #fff;
-  font-size: 13px;
-  font-weight: 500;
-  line-height: 12px;
-}
-
-.header__acc-img {
-  padding: 7px;
-  border-radius: 50%;
-  width: 12px;
-  height: 12px;
-}
-
-.header__acc-img_blue {
-  background-color: #0f4157;
-}
-
-.header__acc-img_dark {
-  background-color: #313131;
-}
-
-.header__auth {
-  display: flex;
-  gap: 30px;
-  margin: 0;
-  padding: 0;
-  list-style-type: none;
-  align-items: center;
-}
-
-.header__auth-reg {
-  margin: 0;
-  padding: 0;
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 16px;
-}
-
-.header__auth-reg:hover {
-  opacity: 0.7;
-  cursor: pointer;
-}
-
-.header__auth-login {
-  margin: 0;
-  color: #000;
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 16px;
-  border-radius: 3px;
-  background: #3ddc84;
-  padding: 8px 20px;
-}
-
-.header__auth-login:hover {
-  cursor: pointer;
-  opacity: 0.8;
-}
-
-.header__burger {
-  background-image: url(../../images/burger-icon.svg);
-  background-repeat: no-repeat;
-  background-color: transparent;
-  width: 44px;
-  height: 44px;
-  background-size: 100%;
-  border: none;
-  display: none;
-  box-sizing: border-box;
-  padding: 0;
-}
-
-.header__burger {
-  opacity: 0.8;
-}
-
-@media (max-width: 768px) {
-  .header__container {
-    padding-top: 15px;
-    padding-bottom: 15px;
-  }
-  .header__links {
-    display: none;
-  }
-
-  .header__acc {
-    display: none;
-  }
-
-  .header__nav {
-    width: calc(100% - 44px);
-    padding-top: 3px;
-    padding-bottom: 3px;
-  }
-
-  .header__burger {
-    display: block;
-  }
-}
-
-@media (max-width: 540px) {
-  .header__container {
-    padding-bottom: 16px;
-  }
-
-  .header__nav {
-    width: calc(100% - 40px);
-    padding-bottom: 2px;
-  }
-
-  .header__burger {
-    width: 40px;
-    height: 43px;
-    display: block;
-  }
-
-  .header__auth {
-    gap: 14px;
-  }
-
-  .header__auth-reg {
-    font-size: 10px;
-  }
-
-  .header__auth-login {
-    font-size: 10px;
-    padding: 5px 12px;
-  }
 }
